+++ .phpstorm.meta.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace PHPSTORM_META;
+
+override(\MarfaTech\Bundle\ApiPlatformBundle\Factory\ApiDtoFactoryDecorate::createApiDto(0), type(0));
+override(\MarfaTech\Bundle\ApiPlatformBundle\Factory\ApiDtoFactory::createApiDto(0), type(0));
+override(\MarfaTech\Bundle\ApiPlatformBundle\Factory\ApiDtoFactory::createApiCollectionDto(0), type(0));
+override(\MarfaTech\Bundle\ApiPlatformBundle\Factory\ApiDtoFactory::createApiDtoByRequest(0), type(0));